+++ packages/fuels-contract/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod contract;
+pub mod logs;
 pub mod predicate;
 pub mod script;
