+++ packages/fuels/src/lib.rs
@@ -62,8 +62,11 @@ pub mod prelude {
     //! use fuels::prelude::*;
     //! ```
 
-    pub use super::contract::contract::{Contract, MultiContractCallHandler};
-    pub use super::contract::predicate::Predicate;
+    pub use super::contract::{
+        contract::{Contract, MultiContractCallHandler},
+        logs::LogDecoder,
+        predicate::Predicate,
+    };
     pub use super::core::constants::*;
     pub use super::core::parameters::*;
     pub use super::core::tx::{Address, AssetId, ContractId};
